Fixing incorrect routes, safe are view, styling issues.
