+++ components/DailyReflection.js
@@ -123,7 +123,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
               top: "50%",
               left: 10,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "PrayerRoom"),
+                completedItem.items.find((item) => item === PRAYER_ROOM_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
@@ -137,7 +137,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
             style={{
               width: 25,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "PrayerRoom"),
+                completedItem.items.find((item) => item === PRAYER_ROOM_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
@@ -208,7 +208,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
               bottom: "50%",
               left: 10,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "VerseOfTheDay"),
+                completedItem.items.find((item) => item === VERSE_OF_THE_DAY_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
@@ -226,7 +226,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
               top: "50%",
               left: 10,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "VerseOfTheDay"),
+                completedItem.items.find((item) => item === VERSE_OF_THE_DAY_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
@@ -240,7 +240,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
             style={{
               width: 25,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "VerseOfTheDay"),
+                completedItem.items.find((item) => item === VERSE_OF_THE_DAY_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
@@ -310,7 +310,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
               bottom: "50%",
               left: 10,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "DevoList"),
+                completedItem.items.find((item) => item === DEVO_LIST_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
@@ -324,7 +324,7 @@ const DailyReflection = ({ completedItems, theme, devoStreak, appStreak }) => {
             style={{
               width: 25,
               backgroundColor: completedItems.some((completedItem) =>
-                completedItem.items.find((item) => item === "DevoList"),
+                completedItem.items.find((item) => item === DEVO_LIST_SCREEN),
               )
                 ? theme === "dark"
                   ? "#a5c9ff"
