+++ hooks/useIsReady.js
@@ -4,7 +4,7 @@ const useIsReady = () => {
   const [isReady, setIsReady] = useState(false);
 
   useEffect(() => {
-    setTimeout(() => setIsReady(true), 500);
+    setTimeout(() => setIsReady(true), 100);
   }, []);
 
   return isReady;
