+++ package.json
@@ -55,7 +55,7 @@
     "react-native-gesture-handler": "~2.16.1",
     "react-native-modal-datetime-picker": "^17.1.0",
     "react-native-paper": "^5.6.0",
-    "react-native-reanimated": "~3.12.0",
+    "react-native-reanimated": "~3.10.1",
     "react-native-safe-area-context": "4.10.1",
     "react-native-size-matters": "^0.4.2",
     "react-native-svg": "^15.3.0",
