+++ routes.js
@@ -1,4 +1,5 @@
 export const HOME_SCREEN = "home";
+export const FOLDER_SCREEN = "folder";
 export const PRAYER_SCREEN = "prayer";
 export const DEVOTIONAL_SCREEN = "devotional";
 export const DEVO_LIST_SCREEN = "devo-list";
